refactor: 컴포넌트 이름 변경 및 코드 정리
