+++ components/button/closeButton/CloseButton.tsx
@@ -1,7 +1,6 @@
 import classNames from "classnames";
 import styles from "./CloseButton.module.scss";
 import Image from "next/image";
-import Button from "../Button";
 
 type buttonProps = {
   className: string;
@@ -10,9 +9,9 @@ type buttonProps = {
 
 function CloseButton({ className, onClick }: buttonProps) {
   return (
-    <Button className={classNames(styles.button, className)} onClick={onClick}>
+    <button className={classNames(styles.button, className)} onClick={onClick}>
       <Image src="/assets/close.svg" alt="닫기 버튼" fill />
-    </Button>
+    </button>
   );
 }
