+++ components/header/Header.tsx
@@ -1,7 +1,7 @@
 import { useEffect, useState } from "react";
 import { getUserProfileById } from "@/lib/api";
 import UserMenu, { UserProfile } from "./userMenu/UserMenu";
-import Container from "@/container/PageContainer";
+import Container from "@/container/Container";
 import styles from "./Header.module.css";
 import Image from "next/image";
 import Link from "next/link";
